This PR implements confirmation box before deletion

+++ lib/services/login_service.dart
@@ -2,14 +2,17 @@ import 'package:cryptokeep/dao/dao.dart';
 import 'package:cryptokeep/exceptions/validation_exceptions.dart';
 import 'package:cryptokeep/models/login_model.dart';
 import 'package:cryptokeep/provider/login_provider.dart';
+import 'package:cryptokeep/utils/app_snackbar.dart';
 import 'package:cryptokeep/utils/validations.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 class LoginService {
   PasswordProvider _provider;
+  BuildContext _context;
 
   LoginService._privateConstructor(BuildContext context) {
+    this._context = context;
     this._provider = Provider.of<PasswordProvider>(context, listen: false);
   }
 
@@ -22,10 +25,36 @@ class LoginService {
 
   void handleDeleteCard(Login object) async {
     this._provider.setIsLoading(true);
-    var response = await DBHelper.instance.deleteByID(object.id);
-    print(response);
-    await Future.delayed(Duration(milliseconds: 350));
+
+    final shouldDelete = await showDialog(
+        context: this._context,
+        builder: (context) {
+          return AlertDialog(
+            title: Text("Delete Login"),
+            content: Text(
+                "Are you sure ? Login '${object.title}' will be deleted permanently"),
+            actions: [
+              MaterialButton(
+                onPressed: () => Navigator.pop(context, false),
+                child: Text("Cancel"),
+              ),
+              MaterialButton(
+                onPressed: () => Navigator.pop(context, true),
+                child: Text("Delete"),
+                color: Colors.redAccent,
+              )
+            ],
+          );
+        });
+
+    if (!shouldDelete) return;
+
+    await DBHelper.instance.deleteByID(object.id);
     this._provider.removeItem(object, isLoading: false);
+
+    await Future.delayed(Duration(milliseconds: 350));
+    AppSnackBar.show(this._context, text: "Login deleted successfully");
+    Navigator.of(this._context).pop();
   }
 
   /// Operations on Logins
